feat: ip links + refactor
